Mod: Separate tests for follow test code convensions
